+++ packages/sqlite3_wasm_build/build.sh
@@ -2,7 +2,7 @@
 set -e
 
 SQLITE_VERSION="2.9.0"
-POWERSYNC_CORE_VERSION="0.4.6"
+POWERSYNC_CORE_VERSION="0.4.8"
 SQLITE_PATH="sqlite3.dart"
 
 if [ -d "$SQLITE_PATH" ]; then